Creating initial server and client files
